+++ package.json
@@ -23,29 +23,29 @@
     "qwik": "qwik"
   },
   "devDependencies": {
-    "@builder.io/qwik": "0.16.2",
-    "@builder.io/qwik-city": "0.1.0-beta9",
+    "@builder.io/qwik": "0.17.5",
+    "@builder.io/qwik-city": "0.1.1",
     "@types/compression": "^1.7.2",
     "@types/cookie": "^0.5.1",
-    "@types/eslint": "8.4.10",
-    "@types/express": "4.17.16",
+    "@types/eslint": "8.21.0",
+    "@types/express": "4.17.17",
     "@types/node": "latest",
     "@types/set-cookie-parser": "^2.4.2",
-    "@typescript-eslint/eslint-plugin": "5.49.0",
-    "@typescript-eslint/parser": "5.49.0",
+    "@typescript-eslint/eslint-plugin": "5.51.0",
+    "@typescript-eslint/parser": "5.51.0",
     "compression": "^1.7.4",
-    "eslint": "8.32.0",
-    "eslint-plugin-qwik": "0.16.2",
+    "eslint": "8.34.0",
+    "eslint-plugin-qwik": "0.17.5",
     "express": "4.18.2",
-    "prettier": "2.8.3",
-    "typescript": "4.9.4",
-    "vite": "4.0.4",
+    "prettier": "2.8.4",
+    "typescript": "4.9.5",
+    "vite": "4.1.1",
     "vite-tsconfig-paths": "4.0.5"
   },
   "dependencies": {
-    "@auth/core": "0.2.5",
+    "@auth/core": "0.4.0",
     "cookie": "^0.5.0",
-    "undici": "5.16.0",
-    "zod": "^3.20.2"
+    "undici": "5.18.0",
+    "zod": "^3.20.6"
   }
 }
